replace import file in help menu
