+++ apps/dotcom/client/src/tla/components/TlaAppMenuGroup/TlaAppMenuGroup.tsx
@@ -1,13 +1,20 @@
 import { useAuth } from '@clerk/clerk-react'
+import { fileOpen } from 'browser-fs-access'
+import { useNavigate } from 'react-router-dom'
 import {
 	ColorSchemeMenu,
 	LanguageMenu,
+	TLDRAW_FILE_EXTENSION,
 	TldrawUiMenuGroup,
 	TldrawUiMenuItem,
 	useDialogs,
 	useMaybeEditor,
 } from 'tldraw'
 import { useOpenUrlAndTrack } from '../../../hooks/useOpenUrlAndTrack'
+import { routes } from '../../../routeDefs'
+import { useMaybeApp } from '../../hooks/useAppState'
+import { useTldrawAppUiEvents } from '../../utils/app-ui-events'
+import { getCurrentEditor } from '../../utils/getCurrentEditor'
 import { defineMessages, useMsg } from '../../utils/i18n'
 import { TlaDebugMenuGroup } from '../TlaDebugMenuGroup'
 import { SubmitFeedbackDialog } from '../dialogs/SubmitFeedbackDialog'
@@ -15,6 +22,7 @@ import { TlaManageCookiesDialog } from '../dialogs/TlaManageCookiesDialog'
 
 const messages = defineMessages({
 	help: { defaultMessage: 'Help' },
+	importFile: { defaultMessage: 'Import file…' },
 	getHelp: { defaultMessage: 'User manual' },
 	legalSummary: { defaultMessage: 'Legal summary' },
 	terms: { defaultMessage: 'Terms of service' },
@@ -22,7 +30,7 @@ const messages = defineMessages({
 	cookiePolicy: { defaultMessage: 'Cookie policy' },
 	manageCookies: { defaultMessage: 'Manage cookies' },
 	about: { defaultMessage: 'About tldraw' },
-	submitFeedback: { defaultMessage: 'Give us feedback' },
+	submitFeedback: { defaultMessage: 'Send feedback' },
 })
 
 export function TlaAppMenuGroup() {
@@ -38,10 +46,11 @@ export function TlaAppMenuGroup() {
 				<LegalSummaryMenuItem />
 				{isSignedIn && <CookieConsentMenuItem />}
 			</TldrawUiMenuGroup>
-			<TldrawUiMenuGroup id="settings">
+			<TldrawUiMenuGroup id="settings-and-rare-actions">
 				<ColorThemeSubmenu />
 				<LanguageMenu />
 				<TlaDebugMenuGroup />
+				<ImportFileActionItem />
 			</TldrawUiMenuGroup>
 		</>
 	)
@@ -125,6 +134,46 @@ function LegalSummaryMenuItem() {
 			readonlyOk
 			onSelect={() => {
 				openAndTrack('https://tldraw.notion.site/legal')
+			}}
+		/>
+	)
+}
+
+function ImportFileActionItem() {
+	const trackEvent = useTldrawAppUiEvents()
+	const app = useMaybeApp()
+
+	const navigate = useNavigate()
+
+	const importFileMsg = useMsg(messages.importFile)
+
+	return (
+		<TldrawUiMenuItem
+			id="about"
+			label={importFileMsg}
+			icon="import"
+			readonlyOk
+			onSelect={async () => {
+				const editor = getCurrentEditor()
+				if (!editor) return
+				if (!app) return
+
+				trackEvent('import-tldr-file', { source: 'account-menu' })
+
+				try {
+					const tldrawFiles = await fileOpen({
+						extensions: [TLDRAW_FILE_EXTENSION],
+						multiple: true,
+						description: 'tldraw project',
+					})
+
+					app.uploadTldrFiles(tldrawFiles, (file) => {
+						navigate(routes.tlaFile(file.id), { state: { mode: 'create' } })
+					})
+				} catch {
+					// user cancelled
+					return
+				}
 			}}
 		/>
 	)
